src/main/java/net/ukr/dreamsicle/validation/ValidCurrencyCodeValidator.java
package net.ukr.dreamsicle.validation;


import com.google.common.base.Strings;

import javax.validation.ConstraintValidator;
import javax.validation.ConstraintValidatorContext;
import java.util.Currency;
import java.util.Set;

public class ValidCurrencyCodeValidator implements ConstraintValidator<ValidCurrencyCode, String> {

    private Boolean isOptional;

    @Override
    public void initialize(ValidCurrencyCode validCurrencyCode) {
        this.isOptional = validCurrencyCode.optional();
    }

    @Override
    public boolean isValid(String value, ConstraintValidatorContext constraintValidatorContext) {

        Set<Currency> currencies = Currency.getAvailableCurrencies();

        return Strings.isNullOrEmpty(value) ? isOptional : currencies.contains(Currency.getInstance(value));
    }
}
